[201911] Verify buffer priority groups for all platforms (#4491)
What is the motivation for this PR?
Verify buffer priority groups for non-Mellanox platforms as well.

Signed-off-by: Stephen Sun stephens@nvidia.com

How did you do it?
To verify:

On Mellanox platform, the lossless buffer profile is applied on PG 3-4 only if the port is admin down
On non-Mellanox platforms, the lossless buffer profile is applied on PG 3-4 of all interfaces regardless of their admin status.

How did you verify/test it?
It was verified on Mellanox platform. An old image without reclaiming buffer is adapted to simulate non-Mellanox platform.