[parallel_run] Run parallel tasks in batches of specified size. (#4509)
Currently the 'parallel_run' API runs the parallel tasks depending on the number of nodes provided in API.

In case of chassis, in some cases, e.g. BGP graceful restart config, the number of parallel tasks can be huge. In T2
topology number of VMs are around 90, this ends up creating 90 parallel tasks which can sometimes overload the
sonic-mgmt server resulting in OOM issues.

Approach
Run parallel tasks in batches of specified size, default is 24.
If a task times out, it remains in workers queue until when the API kills remaining tasks in the end.
If all the tasks in an iteration times out, all those are killed and the test exits with a failure.
If a subset of tasks in an iteration timeout, they continue to be counted as 'tasks running' towards count of concurrent tasks.

What is the motivation for this PR?
Run specific amount of tasks in parallel so that large amount of parallel tasks do not freeze the host

How did you do it?
Run only specified number of tasks in parallel.