Update the mux toggle fixtures to skip non-dualtor topo
The mux toggle fixtures are not just used in the dualtor specific test scripts. Some test scripts
that are generic to both dualtor and non-dualtor topo also use the mux toggle fixtures. These
fixtures need to handle both dualtor and non-dualtor topo.

This change added code of checking topo to the fixtures. When the current topology is
non-dualtor, simply return.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>