Update the mux toggle fixtures to skip non-dualtor topo (#4617)
What is the motivation for this PR?
The mux toggle fixtures are not just used in the dualtor specific test scripts. Some test scripts
that are generic to both dualtor and non-dualtor topo also use the mux toggle fixtures. These
fixtures need to handle both dualtor and non-dualtor topo.

How did you do it?
This change added code checking topo to the fixtures. When the current topology is
non-dualtor, simply return.

How did you verify/test it?
Created a temp script that uses all the touched fixtures.
Tested the temp script on both dualtor and non-dualtor testbeds.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>