Fix the issue of standby_side field always be "lower_tor" in response of mux simulator (#4626)
Because of a logic error, the "standby_side" field in the returned mux status is always "lower_tor".
Originally the standby_side is determined by:

    standby_side = LOWER_TOR if active_side == UPPER_TOR else LOWER_TOR

When "active_side" is not UPPER_TOR, then it must be LOWER_TOR. In this case, the "standby_side" variable
is set with value LOWER_TOR which is same as value of the "active_side" variable.

Fortunately the "standby_side" field is not used by any test or by the simulated y_cable driver, this
issue didn't cause trouble. Anyway, this logic error need to be fixed. The fix is simple, just update
the original code to below:

    standby_side = LOWER_TOR if active_side == UPPER_TOR else UPPER_TOR

Signed-off-by: Xin Wang <xiwang5@microsoft.com>