Reduce unnecessary logs in sanity check Some logs in sanity check may occupy multiple pages. These logs are not necessary for troubleshooting. It would be better to reduce them to make the test log more concise and relevant.
* Reduce logging of content of conditional marks
The conditional marks are defined in yaml file committed in git repository. It's unnecessary to log them
again. If anything is wrong, we can simply check the file in git repository for its content.

* Reduce logging of redis db memory checking command output
The redis db checking will log the full output of redis db cli "client list". This log may occupy multiple
pages. Since this db memory check is pretty stable and never showed issue. It should be safe to
reduce this log too.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>