[Sensors] Check stderr is not None before decode
What is the motivation for this PR?
If running "sensors" succeed, the stderr is None. It will throw out an exception if we decode a None object.

How did you do it?
Check stderr string is not None before decode.

How did you verify/test it?
Run test_sensors.py

Signed-off-by: Kevin(Shengkai) Wang <shengkaiwang@microsoft.com>