Improve mux simulator sanity check (#4640)
What is the motivation for this PR?
When ICMP responder is not running, the standby side may initiate request
switching it self to active. This is by design and could happen just after the
active side of mux simulator is toggled. This would result in unnecessary mux
simulator sanity check failure.

How did you do it?
This change improved the mux simulator sanity check in these ways:
* Do not toggle active side mux ports
* Simply collect mux status of DUTs. Check if the mux status is consistent
between upper_tor and lower_tor.
* Get mux status from mux simulator. Check if the status matches the
actual status on dualtor DUTs. Check the flows in each mux status to
ensure that they are healthy.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>