tests/scp: Improve SCP copy test to work both on old and new image (#4657)
The test initially only worked on newer images but failed on the 201911 image. The 201911 failure was brough up in https://github.com/Azure/sonic-mgmt/issues/4480 and this PR tries to resolve the PR by making sure the test runs on both platforms.

For further context, the failure occurred because pexpect library is not present in python3 but in python2 in the 201911 image. I've run the test on both images in a t0 topology.

What is the motivation for this PR?
Failure of test case in 201911 image.

How did you do it?
Used python2 for 201911 and python3 for newer images.

How did you verify/test it?
Tested change on both 201911 image and a fresh image in a t0 topology.