Make all port channels names padded with one zero
In minigraph_dpg_asic.j2 we are making port channel names be zero padded with upto 2 zeros with 'zfill(2)' - example: PortChannel02
However, in minigraph_dpg.j2, the padded is upto 4 zeros with 'zfill(4)' - example: PortChannel0002

This causes mismatch in the minigraph for the port channels in the host DeviceDataPlaneInfo  vs asic DeviceDataPlaneInfo.

Fix is to use 'zfill(2)' in minigraph_dpg.j2 as well, instead of 'zfill(4)'.