[dualtor_io]: Reduce number of packets sent (#4712)
What is the motivation for this PR?
Recent changes in #4669 mean that switchovers may take longer in failure detection scenarios. Appropriately increase the disruption tolerance to account for this.

Recent changes in #4669 mean that packets take longer to send. Occasionally, this can increase the time taken to send over the 5 minute timeout currently used by the dual ToR sniffer. This can cause packets to be missed from the end of the packet stream, or the sniffer to fail with an exception when the sender attempts to terminate it early:

How did you do it?
Scale the number of packets sent inversely to the interval between packets to ensure that the traffic sending thread always finishes before the sniffer.

How did you verify/test it?
Run several test cases with the new scaling changes, verify that they are able to pass successfully.

Signed-off-by: Lawrence Lee <lawlee@microsoft.com>