[test_configurable_drop_counters] Use vlan interface's MAC address as DMAC
What is the motivation for this PR?
If the ptf sends packets from the port which belongs to the vlan interface, it should use the VLAN interface's MAC as DMAC.
Especially on the dual-tor system, it would use an alternative MAC "00:aa:bb:cc:dd:ee" as the VLAN interface's MAC.

How did you do it?
Use VLAN interface's MAC as DMAC of packets if the ptf tx port is connected to a VLAN port.

How did you verify/test it?
Run test_configurable_drop_counters test case which should pass.

Signed-off-by: Kevin(Shengkai) Wang <shengkaiwang@microsoft.com>