[doc] Add document explaining internal of SONiC test topologies (#4759)
What is the motivation for this PR?
The SONiC test topology is simulated using VMs/dockers and OVS bridges. Details of the topology is complicated and
hard to understand because of lacking documentations. This PR tries to demystify the internals of the SONiC test
topologies by adding diagrams and explanations to the testbed documentations.

How did you do it?
This added document explaining internal of the SONiC test topologies and
fixed some typos.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>