[test gap] announce/withdraw ipv4/ipv6 routes repeatedly to identify resource leak (#4618)
Description of PR

We have observed issues on multiple platform where some of the events were causing resource leak and eventually resulting in TABLE_FULL errors and crashes. This was impacting dataplane traffic.
This test is to simulate multiple route add/remove and check for the system stability.

What is the motivation for this PR?

We have observed issues on multiple platform where some of the events were causing resource leak and eventually resulting in TABLE_FULL errors and crashes. This was impacting dataplane traffic.
This test is to simulate multiple route add/remove and check for the system stability.

How did you do it?

1. Modify the sonic-mgmt/ansible/library/announce_routes.py to support withdraw routes.
2. Add a test case to announce/withdraw routes repeatedly, and check the crm resource on sonic before and after I modify the routes and the test case support different completeness level(debug/basic/confident/thorough) to announce and withdraw routes servel times.

How did you verify/test it?

1. I run the testbed_cli announce_route in order not to influence original function
2. I run the test case using ' pytest --inventory veos_vtb --host-pattern vlab-01 ../tests/stress/test_stress_routes.py --testbed vms-kvm-t0 --testbed_file vtestbed.csv --log-cli-level info --showlocals --assert plain --show-capture no -rav --completeness_level confident --skip_sanity --disable_loganalyzer'

Signed-off-by: Yutong Zhang <yutongzhang@microsoft.com>