[advance-reboot] Fix sad case: don't send traffic through down lag interfaces (#4758)
Fix advance-reboot SAD cases in testcase test_warm_reboot_multi_sad.
The cause for failure was when incorrect participating member interfaces in a LAG are selected.
This issue supposedly started after #3853
As part of PR 3853, some of the SAD case handling was moved out of ptf-tests. Due to this, the port selection for some cases is not done on ptf scripts.
Specifically, presently the traffic is sent through even the LAGs which are brought down. This leads to part of the downstream traffic always being dropped - ultimately leading to warm-up failure.