Fix test_bgp_update_timer failure caused by scapy upgrade in sonic-mgmt docker image
The scapy package in sonic-mgmt docker image was upgraded in PR https://github.com/Azure/sonic-buildimage/pull/8554.
The new version of scapy uses a different way to represent and dissect BGP messages. This caused the failure of
test_bgp_update_time.py.

This PR is to address the compatibility issue of scapy. The test_bgp_update_timer.py script was updated to be
be compatible with both old and new scapy versions.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>