Fix test_bgp_update_timer failure caused by scapy upgrade in sonic-mg docker image (#4774)
What is the motivation for this PR?
The scapy package in sonic-mgmt docker image was upgraded in PR Azure/sonic-buildimage#8554.
The new version of scapy uses a different way to represent and dissect BGP messages. This caused the failure of
test_bgp_update_time.py.

How did you do it?
This PR is to address the compatibility issue of scapy. The test_bgp_update_timer.py script was updated to be
be compatible with both old and new scapy versions.

How did you verify/test it?
Run bgp/test_bgp_update_timer.py using both old and new sonic-mgmt docker image.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>