Merged PR 313: Add test to valid connection graph
This PR is to validate the connection graph for our current str and str2 inventories in the pretest stage. Reason of adding this PR is that onboarding a new testbed to str or str2 need to modify the connection graph files and could have broken them. With this test, any such broken can be caught during pretest while validating the changes.