Merged PR 311: Move all pipelines to use agent pool "nightly2" instead of "nightly"
Move all pipelines to use agent pool "nightly2" instead of "nightly"

Currently all pipelines are using agent pool "nightly" which is a bunch of sonic-mgmt docker
containers running in lab server_15 and server_19. The "nightly" agent pool is a self-hosted agent pool.
Because of potential security issues, 1ES is requesting all teams to migrate from self-hosted agent pools
to 1ES hosted agent pools.

Build team created a new agent pool "nightly2" which is a Azure virtual machine scale set. This new pool
is a bunch of Azure virtual machines that has access to starlab. The number of online virtual machines
can scale out when more pipelines need to run in parallel. When a virtual machine is idle for a specified
time, it will be automatically destroyed. This is called "scale in". The maximum number of virtual machines
in new pool "nightly2" is set to same as "nightly" which is 50. This means that the new "nightly2" pool
also supports at most 50 parallel jobs.

To use the "nightly2" pool, pipeline job must be specified as below:

```
    - job: JobName
      pool: nightly2
      container: sonicdev-microsoft.azurecr.io/docker-sonic-mgmt:sudo
```

The `pool` attribute specifies that agent pool `nightly2` will be used for running this job. Because an agent
in "nightly2" is a virtual machine, what we really needed for running tests is a sonic-mgmt docker container.
Luckily the Azure devops platform natively supports such user scenario. The second attribute `container` is
to specify that all the steps of current job must be executed in a docker container created from the
specified docker image.

Document: https://docs.microsoft.com/en-us/azure/devops/pipelines/process/container-phases?view=azure-devops

After the agent pool migration, each time a job in a pipeline needs to be executed, a docker container
is always created from the specified docker image in a selected (or newly created by scale out) virtual
machine.

**Why use docker image `sonicdev-microsoft.azurecr.io/docker-sonic-mgmt:sudo` instead of
`sonicdev-microsoft.azurecr.io/docker-sonic-mgmt:latest`?**

When the Azure Devops platform tries to create a docker container in virtual machine, it always tries to
add a new user in the docker container for subsequent steps. The default user of the latest
docker-sonic-mgmt image is `AzDevOps` which does not have permission to run "useradd". Creating
docker container will fail like below:

```
2021-11-26T06:00:32.8861823Z ##[command]/usr/bin/docker exec  5f38d4eb72229dcd159d2f17b8a9d3ed04b7c5f990cfe2eea59d660f9347dac7 useradd -m -u 1001 AzDevOps_azpcontainer
2021-11-26T06:00:32.9753674Z useradd: Permission denied.
2021-11-26T06:00:32.9754732Z useradd: cannot lock /etc/passwd; try again later.
```

To workaround the issue, currently we simply created a new `docker-sonic-mgmt:sudo` image to use `root`
as default user based on the existing `docker-sonic-mgmt:latest` image.

The drawback is that the `docker-sonic-mgmt:sudo` image is not always ...