[test_read_mac_metadata] Check and restore the image before doing this test case. (#3854)
In a genernal test, it is expected to use a image(called tested image)
to run all test case in sonic-mgmt.

However, this test case will change the image on DUT and it caused that
the rest of test item will use the image not the same as tested image if the
iteration/image1/image2 do not set properly.

Generally, image1 will be assigned to a stable image and image2 will be
assigned to tested image. It is expected that the image on DUT will be
tested image(i.e. image2) after running this test case. So, add a teardown
process to restore tested image.

Co-authored-by: dennis_chiu <dennis_chiu@edge-core.com>