Check if directory exists before creating it (#4787)
Configlet/test_add_rack.py fails because test doesn't clean up directories that it created during the test run. when run again, test tries to make the same directory again and fails stating that the directory already exists. The test check if the base_dir is present but doesn't verify if the subsequent directories are present or not. It is a simple fix to verify if the subsequent directories are present or not and then proceed to create them.

Co-authored-by: Perumal Venkatesh <pevenkat@cisco.com>