Fix issue of parsing DUT_NAME from vtestbed.csv (#4793)
What is the motivation for this PR?
When we use run_tests.sh to run a test on VS setup and do not support DUT_NAME via the -d argument, test will fail
with error like Failed: One of the specified DUTs ['[vlab-01]'] does not belong to the testbed vms-kvm-t0.
The issue only happen if testbed file vtestbed.csv is used. The problem is with parsing DUT_NAME from
vtestbed.csv file. Currently the code in run_tests.sh parses DUT_NAME as something like '[vlab-01]' and feed it
to pytest command.

How did you do it?
This PR fixed the issue of parsing DUT_NAME by removing the extra `[` and `]`.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>