Improve sanity check to have expected check sequence (#4783)
What is the motivation for this PR?
Currently the sanity check plugin uses inspect.getmembers to find out the check items defined in
checks.py. The problem is that sequence of the check items found from checks.py is not guaranteed
and can cause subtle issue.

The bgp check will check if all BGP sessions are UP. If the DUT was just reloaded and BGP is the first
check item, there is a chance that BGP check will fail. PR #4757
is to address such issue.

A more reliable fix is to have a guaranteed check sequence. The service check item has logic to wait
some time if the DUT was just rebooted or reloaded. With guaranteed check sequence, we can make the
service check always the first sanity check item.

How did you do it?
This PR added a list CHECK_ITEMS in file checks.py. The sanity check plugin runs check items
defined in the list from top to bottom. This can ensure the default sanity check sequence.

When the check sequence needs to be updated, we can update the CHECK_ITEMS list. Or use the sanity
check options --check_items and --post_check_items to affect the default check items.
Please refer to README.md of the sanity check plugin for how to use the options.

How did you verify/test it?
Run a test with sanity check enabled.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>