Cleanup cache for asics as well in the cleanup_cache_for_session fixture
In the cache collected for a multi-asic DUT, there is cache data for asics that
is stored as well, under the <dutName>-asic<asic_id> directory.
This data is not cleaned up as part of cleanup_cache_for_session fixture.

To cleanup the asic specific data, We get the num_asics from the inventory file
(since at this time we don't have the SonicHost object for the DUT),
and if that is more than one, then we pass <dutName>-asic<asic_id> as the zone
parameter to cache.cleanup