Ignore psu's in psu_skip_list in test_master_led
PR#4545 introduced a change in test_master_led to loop through all the PSU's.
However, there could be DUTs where all the PSU's are not present. Thus, this test fails
when checking a PSU that is not present in the DUT.

So, check if the PSU is present in skip_psu_list and if so, then skip checking led on that PSU