Check operational status of ports on fanout when we bring up ports on DUT as part of link flap
In VoQ tests against a T2 VoQ chassis we sometimes see 5-6 seconds delay before
which the ping to the neighbor works after the port on the DUT is brought up as
part of the voq link flap tests.

Currently, when we bring the link up on either the DUT or the fanout, we would
only check if the link is up on the DUT.

What we see is that even though the DUT becomes operationally up, the remote
peer doesn't become operationally up right away at the same time. It seems to take
2-6 seconds more than the DUT port. If the port is not up on the fanout, then the packets from
the DUT is not forwarded to the testbed server, and thus the ping fails.

Fix is to check the link status on the fanout port as well, whenever we bring a port up.