[test gap] Test fdb mac move to identify resource leak  (#4768)
Description of PR
We have observed issues on multiple platform where some of the events were causing resource leak and eventually resulting in TABLE_FULL errors and crashes. This was impacting dataplane traffic.
This test is to test fdb mac move.

What is the motivation for this PR?
We have observed issues on multiple platform where some of the events were causing resource leak and eventually resulting in TABLE_FULL errors and crashes. This was impacting dataplane traffic.
This test is to test fdb mac move.

How did you do it?
I share12k fdb entries with vlan interfaces on dut. Add and remove them repeatedly. And change the start index of the port in order to simulate mac move everytime.

How did you verify/test it?
I test it by running the script 'test_fdb_mac_move.py'

Signed-off-by: Yutong Zhang <yutongzhang@microsoft.com>