Ignore psu's in psu_skip_list in test_master_led (#4818)
What is the motivation for this PR?
PR#4545 introduced a change in test_master_led to loop through all the PSU's. However, there could be DUTs where all the PSU's are not present.

This test fails when checking a PSU that is not present in the DUT.

How did you do it?
Check if the PSU is present in skip_psu_list and if so, then skip checking led on that PSU

How did you verify/test it?
Tested against a DUT where not all PSU's are present, and also a DUT where all PSU's are present.