Merged PR 333: Remove dependency of agent pool for the github merge pipeline
Remove dependency of agent pool for the github merge pipeline

Currently the pipeline for merging github to internal repository depends
on the nightly agent pool. This is indeed unnecessary. This pipeline
should can run on an Azure DevOps agent.

This change removed the dependency of agent pool for the github merge pipeline.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>