Recover bgpd in neighbors parallelly in sanity check.  (#4866)
Description of PR
In sanity check, if the bgp state between dut and neighbors is not "Establish", it will recover the state to "Establish". Before, we recover the state one by one. Now, we recover the state parallelly to save time.

What is the motivation for this PR?
In sanity check, if the bgp state between dut and neighbors is not "Establish", it will recover the state to "Establish". Before, we recover the state one by one. Now, we recover the state parallelly to save time.

How did you do it?
I change for...in... to parallel_run to recover the state parallelly.

How did you verify/test it?
I shutdown the bgp neighbor on dut, and then run sanity check to recover it.

Signed-off-by: Yutong Zhang <yutongzhang@microsoft.com>