Improve infrastructure to support over 100 VMs per server (#4862)
What is the motivation for this PR?
Currently name of VMs running in test server follows the pattern of "VMXXXX". Each "X" is a decimal digit. The first
two digits usually are for server. The last 2 digits are for VMs. Because of this limitation, at most 100 VMs can
be defined for a test server. After the T2 topology is introduced, a single testbed may need more than 100 VMs as
neighbors. We need to get rid of this limitation to support over 100 VMs on single test server.

How did you do it?
This change updated the infrastructure to get rid of this limitation. Now the VM name can have 4 digits or more.

How did you verify/test it?
Run below tasks for both VMs with 4 digits and 5 digits.
* Test run add-topo, remove-topo, deploy-mg
* Test run scripts test_bgp_facts.py, test_nbr_health.py

Signed-off-by: Xin Wang <xiwang5@microsoft.com>