[iface_namingmode]: Fix regex in show interfaces counters test case (#4853)
What is the motivation for this PR?
show interfaces counters was not testing the output lines correctly as the regex match was not matching any
of the output lines. Ideally the regex should match with all output lines except the header lines.

How did you do it?
This fix is to fix the regex expression so that the interfaces list is generated correctly.
Also add an additional assert to ensure interfaces list is generated.
The regex expression will match a string of the format: (Any non-space character followed by any numbers) and this should match all Interface names in default or alias mode.

How did you verify/test it?
Run this test on KVM testbed with T0 topology.