For Ubuntu 18.04, install j2cli 0.3.10 instead of 0.3.12b0
Version 0.3.12b0 is considered a pre-release version on pip, and at any
rate, version 0.3.10 would be installed by default.

More importantly, on 0.3.12b0, the exit code of `j2cli -v` and `j2cli -h` is
1 instead of 0. This breaks the `setup-container.sh` script which
expects an exit code of 0 for `j2cli -v`.

Along with this change, I change the j2cli check in `setup-container.sh`
to have it use `which` instead to just verify the existence of the
binary. That way, even if someone installs 0.3.12b0, it will still work.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>