Merged PR 331: Extend VM id from 2 digits to 3 digits in veos
Extend VM id from 2 digits to 3 digits in veos

PR https://github.com/Azure/sonic-mgmt/pull/4862 improved the infrastructure to support
more than 100 VMs in a same test server.

This is the follow up change to extend the VM id from 2 digits to 3 digits in veos. With
this change, it would be easy to define more VMs in a test server to support testbeds that
may need over 100 VMs on same test server.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>