Modified testcases to check if functionality is supported from platform.json
test_module.py: This testcase checks attributes of DUT and also other cards. Fetching details of other cards from the DUT is not supported.
   -platform.json change:
		- added a get_module_attributes block in platform.json that describes the attributes that cannot be fetched from other cards
   		eg: "get_module_attributes": { "model": false }
   - testcase change:
		- added a method get_module_attributes to fetch the required attribute from the platform.json
		- in testcases of non-supported attributes, added a check that skips the testcase if the fetched attribute is described as
not supported("false") in the platform.json

test_watchdog.py: This testcase suite checks for watchdog arm/disarming for various range of values. Testcase has a autouse fixture to check if the watchdog is disabled after every testcase. But disarming/re-arming with different values is not supported.
    -platform.json change:
		- added watchdog block in platform.json that describes disarm support ("false")
    - testcase change:
		- added get_watchdog method to fetch attribute from platform.json
		- added a check in autouse fixture that skips the testcase by checking disarm support from platform.json

test_sfp.py: This testcase check disabling channel of transceivers. But, disabling channel is not supported
    - platform.json change:
		- added tx_disable_channel attribute for sfps that describes support for disabling transceiver channel
    - testcase change:
		- added a loop that checks to skip every sfp not supported. If all sfps not supported, skips the entire testcase