Enhance test_reporting to support uploading expected test runs data
For nightly tests scheduled by Azure Pipelines, we want to have a clear picture
of when a testbed is expected to trigger a round of testing and upload test
results. Then based on the expected runs data and actual test results uploaded
by testbeds, we can have a metric of percentage of testbeds reliably uploaded
test results.

This change enhanced the test reporting tool to support ingesting expected
test runs data to Kusto.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>