Merged PR 338: Add pipeline for gathering expected test runs of nightly tests
Add pipeline for gathering expected test runs of nightly tests

The nightly test pipelines will upload test results to kusto after
a successful test run. If something went wrong on a testbed, the
nightly test pipeline may exit early and no test results will be
uploaded to Kusto. Just based on the test results already
uploaded to Kusto, it is hard to figure out whether nightly test
is reliably executed on the testbeds. The point is that the
Kusto data does not have information of when a testbed is supposed
to run a testbed and upload test results.

This changed added a script and a pipeline job to inspect the
existing pipelines to find out all the nightly test pipelines.
Then parse their yaml definition file to extract expected test
run schedules. For each of the scheduled test run, upload a
simple record with testbed name and expected nightly test
trigger timestamp to Kusto.

Code change for uploading expected test runs data is in
https://github.com/Azure/sonic-mgmt/pull/4911

Signed-off-by: Xin Wang <xiwang5@microsoft.com>