Fix the logic to fetch effective speed in headroom size calculation
Fix the logic to fetch the effective speed when calculating headroom size in the test case.
Originally, the effective speed is the configured speed.
However, with port auto negotiation supported, we need take more factors into account:
If the port auto negotiation is enabled, the effective speed should be the maximum speed
in the configured advertised speeds list or the supported speeds list
Otherwise, the effective speed is the configured speed.

Signed-off-by: Stephen Sun <stephens@nvidia.com>