Define a class 'NeighborDevice' to extend dict in order to get brief description. (#4946)
Description of PR
The fixture nbrhosts returns a dict, and nbrhosts.values() returns a list of dict with keys 'hosts' and 'conf'. When we need to print the nbrhost, maybe we only need to print the value of 'hosts' to get brief description. So I define a class to extend dict, which will only returns the value of 'hosts'.

What is the motivation for this PR?
The fixture nbrhosts returns a dict, and nbrhosts.values() returns a list of dict with keys 'hosts' and 'conf'. When we need to print the nbrhost, maybe we only need to print the value of 'hosts' to get brief description. So I define a class to extend dict, which will only returns the value of 'hosts'.

How did you do it?
I define a class which has a funtion 'str'

How did you verify/test it?
By running test cases.

Signed-off-by: Yutong Zhang <yutongzhang@microsoft.com>