T2 topology for a VoQ chassis with 2 multi-asic linecards and a supervisor card (#4827)
* T2 topology file for a VoQ chassis with 2 400G multi-asic linecard and supervisor card

The current T2 topology proposed in OC was with 3 single asic linecards. 1 lincard connected
to T3 VM's and the other two connected to T1 VM's. The same routes are advertised across the T1 VM's
to the 2 linecards. This allows for ECMP coverage across fabric in a VoQ Chassis.

We now have multi-asic linecards, and the above ECMP coverage can be accomplished with 2 multi-asic linecards.
So adding this topology for a VoQ Chassis with 2 multi-asic linecards and a supervisor card.

In this topology, we have:

- 3 DUTs - 2 linecards (dut 0, 1) and 1 Supervisor card (dut 2). Each linecard has 2 asics (asic0 and asic1) and each asic has 18 ports
    - asic0 on linecard dut0 connected to 13 T3 VMs - 5 are on 2 port lag and 8 on a single routed port
    - asic1 on linecard dut0 connected to 13 T3 VMs - 5 are on 2 port lag and 8 on a single routed port
    - asic0 on linecard dut1 connected to 13 T1 VMs- 5 are on 2 port lag and 8 on a single routed port
    - asic1 on linecard dut1 connected to 13 T1 VMs - 5 are on 2 port lag and 8 on a single routed port
- ptf ports are numbered 0-71

Since the cards are multi-asic, it provides for coverage of ECMP across fabric - the reason for having 3 linecards in the originally proposed

* Added vips and T1 VM; renamed t2 topology yml file; added min ports t2 topology yml file

Incorporated review comments:

- Added vips to 4 T1 VM's
   - 2 on each asic
      - 1 on LAG VM, and 1 on routed VM
- Removed 400g from the t2 topology filename
- Added T2 topology file with 3 ports per asic
   - 1 eBGP VM on a 2 port LAG
   - 1 eBGP on routed interface