[test_system_health] add dockers check after critical processes restart (#4960)
Signed-off-by: Anton <antonh@nvidia.com>

Summary:
Fixed test issue in which the dockers can be restarted.
If restart some critical processes, as coppmgrd in swss, all the dockers can be restarted. The restart starts after "start" of the process.
To avoid failure of next tests, added a check of critical processes at the end of the "problematic" test.